requirements.txt
flask
psycopg2-binary
lxml
scrapy
ipython
python-dotenv
